+++ myPick-api/src/main/java/com/develop/mypick/api/jwt/TokenProvider.java
@@ -1,4 +1,4 @@
-package com.develop.mypick.common.jwt;
+package com.develop.mypick.api.jwt;
 
 import com.develop.mypick.api.user.service.CustomUserDetailService;
 import com.develop.mypick.common.exception.ErrorCode;
