+++ myPick-common/src/main/java/com/develop/mypick/common/config/ChatGptConfig.java
